Add support for deploying a Systemd service which configures egress dummy interfaces
This assumes that some other component (e.g. cilium) deploys a configmap
which contains the mapping from cluster node to shadow range. The
configmap must be accessible to the Kubelet (can be done e.g. by
deploying a daemonset which mounts the ConfigMap as a volume on the
target nodes)

The script expects the configmap to have a field for each target node.
The contents of each field are expected to be JSON with keys indicating
the interface prefix and the values being objects with fields "base",
"from" and "to" which can be used to construct the shadow egress IPs in
the range by iterating from "from" to "to" and using each value
generated by the iteration as the suffix to "base" (note that this
currently assumes that each egress range is fully contained in a /24).