Ignore `spec.replicas` for autoscaled MachineSets during ArgoCD sync
We need to ignore changes to `spec.replica` when diffing *and* syncing
MachineSets which are managed by a `MachineAutoscaler`.

To do so we 1) dynamically configure `spec.ignoreDifferences` in the
ArgoCD app to ignore `spec.replicas` of each `MachineSet` that's
targeted by a `MachineAutoscaler` and 2) set ArgoCD sync option
`RespectIgnoreDifferences=true` to ensure that we don't overwrite the
machine autoscaler's value in `spec.replicas` if we're applying other
changes.