Print a diagnostic message for OCP < 4.17 or if Espejote isn't installed
We add a new parameter `openshiftVersion` which is expected to be set as
`${dynamic_facts:openshiftVersion}`. We use that parameter to determine
whether to even attempt to configure node disruption policies.

Additionally, we also check `inv.applications` for `espejote` and print
a diagnostic message if we're on OCP < 4.17 or if Espejote isn't
installed.